web-app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-PyMongo==3.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
id==1.5.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
nh3==0.2.21
nltk==3.9.1
openai==1.67.0
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymongo==4.11.1
pyproject_hooks==1.2.0
pytest==8.3.5
python-dotenv==1.0.1
readme_renderer==44.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==76.0.0
sniffio==1.3.1
tqdm==4.67.1
twine==6.1.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
Werkzeug==3.1.3
